view product

+++ NhuYShop/Model/ProductModel.cs
@@ -8,8 +8,8 @@ namespace NhuYShop.Model
     {
         public int PRODUCTID { get; set; }
         public string PRODUCTNAME { get; set; }
-        public int VALUE1 { get; set; }
-        public int VALUE2 { get; set; }
-        public int VALUE3 { get; set; }
+        public double VALUE1 { get; set; }
+        public double VALUE2 { get; set; }
+        public double VALUE3 { get; set; }
     }
 }
